This is an example refactor of the code found at https://github.com/tuzz/range_splitter/blob/master/lib/range_splitter.rb
However, I could not unsee the refactor attempt made here
https://github.com/andhapp/range_splitter/blob/refactoring/lib/range_splitter.rb

So, being pragmatic (and rather short of time) I built on top of that refactor
which I thought was a pretty good start.

The basic premise of this commit is that strings are also a valid range
so they should join the party. Ranges can now be split over any utf-8 char.

The groundwork has been set for adding other types of ranges using a Factory
to build the rage partition..

I added some type checks and guards to ensure correct behavior and checked
exception.

Updated split documentation and updated readme.